// frontend/src/app/api/models/cv.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CvEducation } from '../models/cv-education';
import { CvExperience } from '../models/cv-experience';
import { CvLanguage } from '../models/cv-language';
import { CvPerson } from '../models/cv-person';
import { CvProject } from '../models/cv-project';
import { CvSkillGroup } from '../models/cv-skill-group';

/**
 * Lebenslauf. end fehlt = bis heute. updated_at wird beim Speichern ignoriert. Solange kein Lebenslauf gespeichert ist, liefert GET eine leere Struktur (person.name leer, Listen leer, ohne updated_at).
 */
export interface Cv {
  education: Array<CvEducation>;
  experience: Array<CvExperience>;
  languages: Array<CvLanguage>;
  person: CvPerson;
  projects: Array<CvProject>;
  skills: Array<CvSkillGroup>;
  summary?: string;
  updated_at?: string;
}